leroyMerlin/requirements.txt
requests~=2.28.1
pandas~=1.4.4
openpyxl~=3.0.10
python-dotenv~=0.21.0
PySimpleGUI~=4.60.3
fake_useragent